Src directory refactor continued
