+++ src/schemaRendering/schemaElements/Checkbox.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import FormElementWrapper from "./FormElementWrapper"
+import FormElementWrapper from "../utils/FormElementWrapper"
 
 function Checkbox(props) {
   const [isChecked, setIsChecked] = useState(false);
