+++ src/schemaRendering/schemaElements/RadioGroup.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import FormElementWrapper from "./FormElementWrapper"
+import FormElementWrapper from "../utils/FormElementWrapper"
 
 function RadioGroup(props) {
   const [value, setValue] = useState("");
