+++ src/schemaRendering/schemaElements/Unit.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import FormElementWrapper from "./FormElementWrapper";
+import FormElementWrapper from "../utils/FormElementWrapper";
 
 function Unit(props) {
   const [value, setValue] = useState("");
