+++ src/schemaRendering/utils/Label.js
